Seperation of concerns in context
