+++ src/Components/Pages/Visualizer/Components/Ancillary/Dropdown.js
@@ -1,9 +1,7 @@
-import React, { useState } from "react";
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import FormHelperText from "@mui/material/FormHelperText";
 import FormControl from "@mui/material/FormControl";
+import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
+import React, { useState } from "react";
 
 // Default Array for before loading 
 const defaultArray = [1];
